requirements.txt
gdown
gradio
numpy
Pillow
torch
torchvision
tqdm
tensorboardX==2.5
matplotlib
faiss-cpu==1.7.2
protobuf<3.21.0